requirements.txt
asgiref==3.7.2
Django==4.2.3
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
future==0.18.3
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
serializers==0.2.4
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
djchoices